requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Database & Data Loading
sqlalchemy>=2.0.0
pyarrow>=12.0.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ML Utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Experiment Tracking (optional)
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
